[DataCollection] fixed an upload bug of images.
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@40588 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/DataCollection/classes/class.ilDataCollectionRecord.php
@@ -554,8 +554,10 @@ class ilDataCollectionRecord
       */
     public function deleteMob($obj_id)
     {
-        $mob = new ilObjMediaObject($obj_id);
-        $mob->delete();
+        if(ilObject2::_lookupObjId($obj_id)){
+            $mob = new ilObjMediaObject($obj_id);
+            $mob->delete();
+        }
     }
 	
 	/*
